+++ atest/_resources/fixtures/api/NodeLabels.py
@@ -1,4 +1,4 @@
-from ipyforcegraph.forcegraph import WIDGET_CLASS
+from ipyforcegraph.graphs import WIDGET_CLASS
 from ipyforcegraph import behaviors as B
 fg = WIDGET_CLASS(behaviors=[B.NodeLabels(column_name="id")])
 fg.source.nodes = [{"id": "hello world"}]